move Citizens to new npcs package
