+++ src/main/java/org/betonquest/betonquest/compatibility/holograms/HologramProvider.java
@@ -6,7 +6,7 @@ import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.compatibility.Compatibility;
 import org.betonquest.betonquest.compatibility.Integrator;
-import org.betonquest.betonquest.compatibility.citizens.CitizensHologramLoop;
+import org.betonquest.betonquest.compatibility.npcs.citizens.CitizensHologramLoop;
 import org.betonquest.betonquest.exceptions.HookException;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.bukkit.Bukkit;
